docs: Split examples into separate folders using a main README.md
