+++ packages/flutter_gamepads/example/basic/pubspec.yaml
@@ -1,0 +1,22 @@
+name: flutter_gamepads_flame_basic
+description: A simple example project showcasing the flutter_gamepads plugin.
+resolution: workspace
+publish_to: 'none'
+
+version: 0.1.0
+
+environment:
+  sdk: ">=3.9.0 <4.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flutter_gamepads: ^0.1.12
+
+dev_dependencies:
+  flame_lint: ^1.4.1
+  flutter_test:
+    sdk: flutter
+
+flutter:
+  uses-material-design: true
