+++ packages/flutter_gamepads/example/flame/analysis_options.yaml
@@ -1,1 +1,4 @@
+analyzer:
+  exclude:
+    - build/**
 include: package:flame_lint/analysis_options.yaml
