+++ packages/flutter_gamepads/example/flame/assets/images/power_up.png
