+++ packages/flutter_gamepads/example/flame/assets/images/spaceship.png
