+++ packages/flutter_gamepads/example/flame/lib/overlays/help_overlay.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_gamepads_example/flame_example/game.dart';
-import 'package:flutter_gamepads_example/flame_example/overlays/overlay_dialog_backdrop.dart';
-import 'package:flutter_gamepads_example/flame_example/overlays/overlays.dart';
+import 'package:flutter_gamepads_flame_example/game.dart';
+import 'package:flutter_gamepads_flame_example/overlays/overlay_dialog_backdrop.dart';
+import 'package:flutter_gamepads_flame_example/overlays/overlays.dart';
 
 class HelpOverlay extends StatelessWidget {
   final MyGame game;
