+++ packages/flutter_gamepads/example/flame/lib/overlays/statusbar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_gamepads_example/flame_example/game.dart';
-import 'package:flutter_gamepads_example/flame_example/overlays/overlays.dart';
+import 'package:flutter_gamepads_flame_example/game.dart';
+import 'package:flutter_gamepads_flame_example/overlays/overlays.dart';
 
 class StatusBarOverlay extends StatefulWidget {
   final MyGame game;
