+++ packages/flutter_gamepads/example/flame/lib/state/game_state.dart
