+++ packages/flutter_gamepads/example/flame/lib/theme.dart
