+++ packages/flutter_gamepads/example/flame/pubspec.yaml
@@ -1,4 +1,4 @@
-name: flutter_gamepads_example
+name: flutter_gamepads_flame_example
 description: A simple example project showcasing the flutter_gamepads plugin.
 resolution: workspace
 publish_to: 'none'
