+++ packages/flutter_gamepads/example/flutter/analysis_options.yaml
@@ -1,0 +1,4 @@
+analyzer:
+  exclude:
+    - build/**
+include: package:flame_lint/analysis_options.yaml
