+++ packages/flutter_gamepads/example/flutter/lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gamepads/flutter_gamepads.dart';
-import 'package:flutter_gamepads_example/flutter_example/pages/game_page.dart';
-import 'package:flutter_gamepads_example/flutter_example/pages/home_page.dart';
-import 'package:flutter_gamepads_example/flutter_example/pages/settings_page.dart';
-import 'package:flutter_gamepads_example/flutter_example/theme.dart';
+import 'package:flutter_gamepads_flutter_example/pages/game_page.dart';
+import 'package:flutter_gamepads_flutter_example/pages/home_page.dart';
+import 'package:flutter_gamepads_flutter_example/pages/settings_page.dart';
+import 'package:flutter_gamepads_flutter_example/theme.dart';
 
 void main() {
   runApp(const MyFlutterApp());
