+++ packages/flutter_gamepads/example/flutter/lib/pages/slider_with_gamepad_support.dart
