+++ packages/flutter_gamepads/example/flutter/pubspec.yaml
@@ -1,0 +1,24 @@
+name: flutter_gamepads_flutter_example
+description: A simple example project showcasing the flutter_gamepads plugin.
+resolution: workspace
+publish_to: 'none'
+
+version: 0.1.0
+
+environment:
+  sdk: ">=3.9.0 <4.0.0"
+
+dependencies:
+  collection: ^1.19.1
+  flutter:
+    sdk: flutter
+  flutter_gamepads: ^0.1.12
+  gamepads: ^0.1.11
+
+dev_dependencies:
+  flame_lint: ^1.4.1
+  flutter_test:
+    sdk: flutter
+
+flutter:
+  uses-material-design: true
